Add files via upload
Developed a camera-based theft detection system using the ESP32-CAM microcontroller integrated with a Telegram bot. The system uses a PIR sensor to detect motion, upon which it captures an image of the intruder. Instead of storing the image locally, it sends the captured image instantly to a Telegram bot, enabling real-time alerts and remote monitoring. This setup enhances security by providing immediate notifications to the user in case of unauthorized access.